+++ classes/item.rb
@@ -23,7 +23,6 @@ class Item
 
   def add_label(label)
     @label = label
-    label.add_item(label)
   end
 
   def add_author(author)
@@ -43,4 +42,3 @@ class Item
     false
   end
 end
-
